Modificado todo layout, tive de mudar a estrutura das duvidas para encaixar o botão de show More/lass
